refactor: move provider launchers from domain to infrastructure
Extract launcher implementations (Subprocess, Docker, Container, HTTP) from
domain/services/provider_launcher/ into infrastructure/launchers/.
Old paths become deprecation shims re-exporting from the new location.

Add domain contracts: IProviderLauncher (launcher.py), ILock (lock.py).
Update IProviderRepository and ProviderRuntime protocol.

This enforces the DDD boundary: domain holds abstractions, infrastructure
holds implementations.
