+++ src/mcp_hangar/domain/services/__init__.py
@@ -1,27 +1,53 @@
 """Domain services - interfaces for infrastructure operations."""
 
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
 # Re-export exception from canonical location for convenience
 from ..exceptions import ProviderStartError
+from ..contracts.launcher import IProviderLauncher, LaunchResult
 from .audit_service import AuditService
 from .error_diagnostics import collect_startup_diagnostics, get_suggestion_for_error
 from .image_builder import BuildConfig, get_image_builder, ImageBuilder
-from .provider_launcher import (
-    ContainerConfig,
-    ContainerLauncher,
-    DockerLauncher,
-    get_launcher,
-    HttpLauncher,
-    ProviderLauncher,
-    SubprocessLauncher,
-)
 from .tool_access_resolver import (
     get_tool_access_resolver,
     reset_tool_access_resolver,
     ToolAccessResolver,
 )
 
+if TYPE_CHECKING:
+    from mcp_hangar.infrastructure.launchers import (
+        ContainerConfig,
+        ContainerLauncher,
+        DockerLauncher,
+        HttpLauncher,
+        ProviderLauncher,
+        SubprocessLauncher,
+        get_launcher,
+    )
+
+
+def __getattr__(name: str) -> object:
+    if name in {
+        "ContainerConfig",
+        "ContainerLauncher",
+        "DockerLauncher",
+        "get_launcher",
+        "HttpLauncher",
+        "ProviderLauncher",
+        "SubprocessLauncher",
+    }:
+        from . import provider_launcher as launcher_module
+
+        return getattr(launcher_module, name)
+    raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
+
+
 __all__ = [
     "AuditService",
+    "IProviderLauncher",
+    "LaunchResult",
     "ProviderLauncher",
     "SubprocessLauncher",
     "DockerLauncher",
