+++ src/mcp_hangar/domain/services/provider_launcher/factory.py
@@ -1,34 +1,16 @@
-"""Factory function for provider launchers."""
+"""Deprecated provider launcher import shim."""
 
-from .base import ProviderLauncher
-from .container import ContainerLauncher
-from .http import HttpLauncher
-from .subprocess import SubprocessLauncher
+from __future__ import annotations
 
+import warnings
 
-def get_launcher(mode: str) -> ProviderLauncher:
-    """
-    Factory function to get the appropriate launcher for a mode.
+from mcp_hangar.infrastructure.launchers.factory import get_launcher
 
-    Args:
-        mode: Provider mode (subprocess, docker, container, podman, remote)
+warnings.warn(
+    "mcp_hangar.domain.services.provider_launcher.factory is deprecated; import from "
+    "mcp_hangar.infrastructure.launchers.factory instead.",
+    DeprecationWarning,
+    stacklevel=2,
+)
 
-    Returns:
-        Appropriate launcher instance
-
-    Raises:
-        ValueError: If mode is not supported
-    """
-    launchers = {
-        "subprocess": SubprocessLauncher,
-        "docker": lambda: ContainerLauncher(runtime="auto"),  # Use ContainerLauncher with auto-detection
-        "container": lambda: ContainerLauncher(runtime="auto"),
-        "podman": lambda: ContainerLauncher(runtime="podman"),
-        "remote": HttpLauncher,
-    }
-
-    launcher_factory = launchers.get(mode)
-    if not launcher_factory:
-        raise ValueError(f"unsupported_mode: {mode}")
-
-    return launcher_factory() if callable(launcher_factory) else launcher_factory
+__all__ = ["get_launcher"]
