+++ src/mcp_hangar/infrastructure/__init__.py
@@ -24,6 +24,15 @@ from .event_store import (
     InMemoryEventStore,
     StoredEvent,
 )
+from .launchers import (
+    ContainerConfig,
+    ContainerLauncher,
+    DockerLauncher,
+    get_launcher,
+    HttpLauncher,
+    ProviderLauncher,
+    SubprocessLauncher,
+)
 from .lock_hierarchy import (
     clear_thread_locks,
     get_current_thread_locks,
@@ -75,6 +84,14 @@ __all__ = [
     "EventStoreSnapshot",
     "ConcurrencyError",
     "get_event_store",
+    # Launchers
+    "ProviderLauncher",
+    "SubprocessLauncher",
+    "DockerLauncher",
+    "ContainerLauncher",
+    "ContainerConfig",
+    "HttpLauncher",
+    "get_launcher",
     # Event Sourced Repository
     "EventSourcedProviderRepository",
     "ProviderConfigStore",
