+++ src/mcp_hangar/infrastructure/launchers/base.py
@@ -1,0 +1,11 @@
+"""Base provider launcher interface."""
+
+from __future__ import annotations
+
+
+class ProviderLauncher:
+    """Infrastructure base class for provider launchers."""
+
+    def stop(self, provider_id: str) -> None:
+        """Stop a provider previously launched by this launcher."""
+        _ = provider_id
